+++ packages/cli/src/util/paths.ts
@@ -16,7 +16,7 @@ import * as Either from 'fp-ts/lib/Either';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { get } from 'lodash';
 // @ts-ignore
-import { parse } from 'uri-template';
+import { URI } from 'uri-template-lite';
 
 export function createExamplePath(operation: IHttpOperation): Either.Either<Error, string> {
   return pipe(
@@ -29,7 +29,7 @@ export function createExamplePath(operation: IHttpOperation): Either.Either<Erro
         })
       );
     }),
-    Either.map(({ template, values }) => parse(template).expand(values))
+    Either.map(({ template, values }) => URI.expand(template, values))
   );
 }