+++ packages/http/src/validator/deserializers/index.ts
@@ -7,6 +7,8 @@ import {
   FormStyleDeserializer,
   SimpleStyleDeserializer,
 } from './style';
+import { LabelStyleDeserializer } from './style/label';
+import { MatrixStyleDeserializer } from './style/matrix';
 
 export const header = new HttpParamDeserializerRegistry([new SimpleStyleDeserializer()]);
 
@@ -16,6 +18,12 @@ export const query = new HttpParamDeserializerRegistry([
   new DelimitedStyleDeserializer('|', HttpParamStyles.PipeDelimited),
   new DelimitedStyleDeserializer(',', HttpParamStyles.CommaDelimited),
   new DeepObjectStyleDeserializer(),
+]);
+
+export const path = new HttpParamDeserializerRegistry([
+  new SimpleStyleDeserializer(),
+  new LabelStyleDeserializer(),
+  new MatrixStyleDeserializer(),
 ]);
 
 export const body = query;