+++ packages/http/src/validator/validators/__tests__/path.spec.ts
@@ -1,10 +1,11 @@
 import { HttpParamStyles, IHttpQueryParam } from '@stoplight/types';
-import { query as registry } from '../../deserializers';
-import { HttpQueryValidator } from '../query';
+import { path as registry } from '../../deserializers';
+import { HttpPathValidator } from '../path';
 import * as validateAgainstSchemaModule from '../utils';
+import { IHttpPathParam } from '@stoplight/types/dist';
 
-describe('HttpQueryValidator', () => {
-  const httpQueryValidator = new HttpQueryValidator(registry, 'query');
+describe('HttpPathValidator', () => {
+  const httpPathValidator = new HttpPathValidator(registry, 'path');
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -14,41 +15,46 @@ describe('HttpQueryValidator', () => {
 
   describe('validate()', () => {
     describe('spec is present', () => {
-      describe('query param is not present', () => {
+      describe('path param is not present', () => {
         describe('spec defines it as required', () => {
           it('returns validation error', () => {
             expect(
-              httpQueryValidator.validate({}, [{ name: 'aParam', style: HttpParamStyles.Form, required: true }]),
-            ).toMatchSnapshot();
+              httpPathValidator.validate({}, [{ name: 'aParam', style: HttpParamStyles.Simple, required: true }]),
+            ).toEqual([{
+            code: 'required',
+              message: 'should have required property \'aparam\'',
+              path: ['path'],
+              severity: 0,
+            }]);
           });
         });
       });
 
-      describe('query param is present', () => {
+      describe('path param is present', () => {
         describe('schema is present', () => {
           describe('deserializer not available', () => {
             it('omits schema validation', () => {
               jest.spyOn(registry, 'get').mockReturnValueOnce(undefined);
-              const param: IHttpQueryParam = {
+              const param: IHttpPathParam = {
                 name: 'param',
-                style: HttpParamStyles.Form,
+                style: HttpParamStyles.Simple,
                 schema: { type: 'number' },
               };
 
-              expect(httpQueryValidator.validate({ param: 'abc' }, [param])).toEqual([]);
+              expect(httpPathValidator.validate({ param: 'abc' }, [param])).toEqual([]);
 
               expect(validateAgainstSchemaModule.validateAgainstSchema).toReturnWith([]);
             });
           });
 
           describe('deserializer is available', () => {
-            describe('query param is valid', () => {
+            describe('path param is valid', () => {
               it('validates positively against schema', () => {
                 expect(
-                  httpQueryValidator.validate({ param: 'abc' }, [
+                  httpPathValidator.validate({ param: 'abc' }, [
                     {
                       name: 'param',
-                      style: HttpParamStyles.Form,
+                      style: HttpParamStyles.Simple,
                       schema: { type: 'string' },
                     },
                   ]),
@@ -63,10 +69,10 @@ describe('HttpQueryValidator', () => {
         describe('schema was not provided', () => {
           it('omits schema validation', () => {
             expect(
-              httpQueryValidator.validate({ param: 'abc' }, [
+              httpPathValidator.validate({ param: 'abc' }, [
                 {
                   name: 'param',
-                  style: HttpParamStyles.Form,
+                  style: HttpParamStyles.Simple,
                 },
               ]),
             ).toEqual([]);
@@ -78,14 +84,19 @@ describe('HttpQueryValidator', () => {
         describe('deprecated flag is set', () => {
           it('returns deprecation warning', () => {
             expect(
-              httpQueryValidator.validate({ param: 'abc' }, [
+              httpPathValidator.validate({ param: 'abc' }, [
                 {
                   name: 'param',
                   deprecated: true,
-                  style: HttpParamStyles.Form,
+                  style: HttpParamStyles.Simple,
                 },
               ]),
-            ).toMatchSnapshot();
+            ).toEqual([{
+              code: 'deprecated',
+              message: 'Path param param is deprecated',
+              path: ['path', 'param'],
+              severity: 1,
+            }]);
           });
         });
       });