+++ packages/http/src/validator/validators/path.ts
@@ -1,0 +1,18 @@
+import { HttpParamStyles, IHttpPathParam } from '@stoplight/types';
+import { IPrismDiagnostic } from '@stoplight/prism-core';
+import { IHttpNameValue } from '../../types';
+import { IHttpParamDeserializerRegistry } from '../deserializers/types';
+import { HttpParamsValidator } from './params';
+
+export class HttpPathValidator extends HttpParamsValidator<IHttpNameValue> {
+  constructor(
+    registry: IHttpParamDeserializerRegistry<IHttpNameValue>,
+    prefix: string,
+    style: HttpParamStyles = HttpParamStyles.Simple,
+  ) {
+    super(registry, prefix, style);
+  }
+  public validate(target: IHttpNameValue, specs: IHttpPathParam[]): IPrismDiagnostic[] {
+    return super.validate(target, specs);
+  }
+}